Removes ManifestGenerator
As all AsicEAttachables must return bytes, it is better to have this responsibility within Manifest itself.
