+++ Digipost.Signature.Api.Client.Core.Tests/Asice/AsiceManifest/ManifestTests.cs
@@ -1,4 +1,7 @@
-﻿using Digipost.Signature.Api.Client.Core.Asice.AsiceManifest;
+﻿using System.Text;
+using Digipost.Signature.Api.Client.Core.Asice;
+using Digipost.Signature.Api.Client.Core.Asice.AsiceManifest;
+using Digipost.Signature.Api.Client.Core.Asice.DataTransferObjects;
 using Digipost.Signature.Api.Client.Core.Tests.Utilities;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
@@ -28,6 +31,26 @@ namespace Digipost.Signature.Api.Client.Core.Tests.Asice.AsiceManifest
                 Assert.AreEqual(sender, manifest.Sender);
                 Assert.AreEqual(document, manifest.Document);
                 Assert.AreEqual(signers, manifest.Signers);
+            }
+        }
+
+        [TestClass]
+        public class BytesMethod : ManifestTests
+        {
+            [TestMethod]
+            public void SuccessfulManifestToBytes()
+            {
+                //Arrange
+                var manifest = DomainUtility.GetManifest();
+                var manifestDataTranferObject = DataTransferObjectConverter.ToDataTransferObject(manifest);
+                var expectedResult = SerializeUtility.Serialize(manifestDataTranferObject);
+
+                //Act
+                var bytes = manifest.Bytes;
+                var actualResult = Encoding.UTF8.GetString(bytes);
+
+                //Assert
+                Assert.AreEqual(expectedResult, actualResult);
             }
         }
     }
